The update dialog has been fixed.
The updateDialog had a bug where, when trying to modify a resource for the first time, the confirmation button on the last step was disabled, as if the edit form had not been touched. Now, any change can be confirmed, and if no modifications are made compared to the original, confirmation is not allowed.
The issue with editing the price in the product resource was also fixed. The problem was that the original price entered into the form was a string, while the edited version was a number, causing an incorrect comparison.